Tests para NumericPhenotype

+++ src/main/java/ar/edu/unlp/pasae/tp_integrador/entities/NumericPhenotype.java
@@ -19,7 +19,7 @@ public class NumericPhenotype extends Phenotype {
   @Override
   public Boolean validate(String value) {
     try {
-      Integer.parseInt(value);
+      Float.parseFloat(value);
       return true;
     } catch (Exception NumberFormatException) {
       return false;
